web/flasktest/requirements.txt
azure-common==1.1.27
azure-core==1.21.1
azure-identity==1.7.1
azure-keyvault-secrets==4.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
cryptography==36.0.1
Flask==2.0.2
idna==3.3
isodate==0.6.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
msal==1.16.0
msal-extensions==0.3.1
msrest==0.6.21
oauthlib==3.1.1
portalocker==2.3.2
pycparser==2.21
PyJWT==2.3.0
python-dotenv==0.19.2
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.2
